+++ client_test.go
@@ -25,15 +25,16 @@ import (
 	"github.com/pilosa/pilosa"
 	"github.com/pilosa/pilosa/internal"
 	"github.com/pilosa/pilosa/pql"
+	"github.com/pilosa/pilosa/test"
 )
 
-func createCluster(c *pilosa.Cluster) ([]*Server, []*Holder) {
+func createCluster(c *pilosa.Cluster) ([]*test.Server, []*test.Holder) {
 	numNodes := len(c.Nodes)
-	hldr := make([]*Holder, numNodes)
-	server := make([]*Server, numNodes)
+	hldr := make([]*test.Holder, numNodes)
+	server := make([]*test.Server, numNodes)
 	for i := 0; i < numNodes; i++ {
-		hldr[i] = MustOpenHolder()
-		server[i] = NewServer()
+		hldr[i] = test.MustOpenHolder()
+		server[i] = test.NewServer()
 		server[i].Handler.Host = server[i].Host()
 		server[i].Handler.Cluster = c
 		server[i].Handler.Cluster.Nodes[i].Host = server[i].Host()
@@ -44,7 +45,7 @@ func createCluster(c *pilosa.Cluster) ([]*Server, []*Holder) {
 
 // Test distributed TopN Row count across 3 nodes.
 func TestClient_MultiNode(t *testing.T) {
-	cluster := NewCluster(3)
+	cluster := test.NewCluster(3)
 	s, hldr := createCluster(cluster)
 
 	for i := 0; i < len(cluster.Nodes); i++ {
@@ -197,17 +198,17 @@ func TestClient_MultiNode(t *testing.T) {
 
 // Ensure client can bulk import data.
 func TestClient_Import(t *testing.T) {
-	hldr := MustOpenHolder()
+	hldr := test.MustOpenHolder()
 	defer hldr.Close()
 
 	// Load bitmap into cache to ensure cache gets updated.
 	f := hldr.MustCreateFragmentIfNotExists("i", "f", pilosa.ViewStandard, 0)
 	f.Row(0)
 
-	s := NewServer()
+	s := test.NewServer()
 	defer s.Close()
 	s.Handler.Host = s.Host()
-	s.Handler.Cluster = NewCluster(1)
+	s.Handler.Cluster = test.NewCluster(1)
 	s.Handler.Cluster.Nodes[0].Host = s.Host()
 	s.Handler.Holder = hldr.Holder
 
@@ -232,7 +233,7 @@ func TestClient_Import(t *testing.T) {
 
 // Ensure client can bulk import data to an inverse frame.
 func TestClient_ImportInverseEnabled(t *testing.T) {
-	hldr := MustOpenHolder()
+	hldr := test.MustOpenHolder()
 	defer hldr.Close()
 
 	idx := hldr.MustCreateIndexIfNotExists("i", pilosa.IndexOptions{})
@@ -255,10 +256,10 @@ func TestClient_ImportInverseEnabled(t *testing.T) {
 	// Load bitmap into cache to ensure cache gets updated.
 	f.Row(0)
 
-	s := NewServer()
+	s := test.NewServer()
 	defer s.Close()
 	s.Handler.Host = s.Host()
-	s.Handler.Cluster = NewCluster(1)
+	s.Handler.Cluster = test.NewCluster(1)
 	s.Handler.Cluster.Nodes[0].Host = s.Host()
 	s.Handler.Holder = hldr.Holder
 
@@ -287,7 +288,7 @@ func TestClient_ImportInverseEnabled(t *testing.T) {
 
 // Ensure client backup and restore a frame.
 func TestClient_BackupRestore(t *testing.T) {
-	hldr := MustOpenHolder()
+	hldr := test.MustOpenHolder()
 	defer hldr.Close()
 
 	hldr.MustCreateFragmentIfNotExists("i", "f", pilosa.ViewStandard, 0).MustSetBits(100, 1, 2, 3, SliceWidth-1)
@@ -295,10 +296,10 @@ func TestClient_BackupRestore(t *testing.T) {
 	hldr.MustCreateFragmentIfNotExists("i", "f", pilosa.ViewStandard, 5).MustSetBits(100, (5*SliceWidth)+1)
 	hldr.MustCreateFragmentIfNotExists("i", "f", pilosa.ViewStandard, 0).MustSetBits(200, 20000)
 
-	s := NewServer()
+	s := test.NewServer()
 	defer s.Close()
 	s.Handler.Host = s.Host()
-	s.Handler.Cluster = NewCluster(1)
+	s.Handler.Cluster = test.NewCluster(1)
 	s.Handler.Cluster.Nodes[0].Host = s.Host()
 	s.Handler.Holder = hldr.Holder
 
@@ -335,7 +336,7 @@ func TestClient_BackupRestore(t *testing.T) {
 
 // Ensure client backup and restore a frame with inverse view.
 func TestClient_BackupInverseView(t *testing.T) {
-	hldr := MustOpenHolder()
+	hldr := test.MustOpenHolder()
 	defer hldr.Close()
 
 	idx := hldr.MustCreateIndexIfNotExists("i", pilosa.IndexOptions{})
@@ -360,10 +361,10 @@ func TestClient_BackupInverseView(t *testing.T) {
 	f.SetBit(100, 3)
 	f.SetBit(100, SliceWidth-1)
 
-	s := NewServer()
+	s := test.NewServer()
 	defer s.Close()
 	s.Handler.Host = s.Host()
-	s.Handler.Cluster = NewCluster(1)
+	s.Handler.Cluster = test.NewCluster(1)
 	s.Handler.Cluster.Nodes[0].Host = s.Host()
 	s.Handler.Holder = hldr.Holder
 
@@ -392,15 +393,15 @@ func TestClient_BackupInverseView(t *testing.T) {
 
 // backup returns error with invalid view
 func TestClient_BackupInvalidView(t *testing.T) {
-	hldr := MustOpenHolder()
+	hldr := test.MustOpenHolder()
 	defer hldr.Close()
 
 	hldr.MustCreateFragmentIfNotExists("i", "f", pilosa.ViewStandard, 0).MustSetBits(100, 1, 2, 3, SliceWidth-1)
 
-	s := NewServer()
+	s := test.NewServer()
 	defer s.Close()
 	s.Handler.Host = s.Host()
-	s.Handler.Cluster = NewCluster(1)
+	s.Handler.Cluster = test.NewCluster(1)
 	s.Handler.Cluster.Nodes[0].Host = s.Host()
 	s.Handler.Holder = hldr.Holder
 
@@ -416,7 +417,7 @@ func TestClient_BackupInvalidView(t *testing.T) {
 
 // Ensure client can retrieve a list of all checksums for blocks in a fragment.
 func TestClient_FragmentBlocks(t *testing.T) {
-	hldr := MustOpenHolder()
+	hldr := test.MustOpenHolder()
 	defer hldr.Close()
 
 	// Set two bits on blocks 0 & 3.
@@ -426,10 +427,10 @@ func TestClient_FragmentBlocks(t *testing.T) {
 	// Set a bit on a different slice.
 	hldr.MustCreateFragmentIfNotExists("i", "f", pilosa.ViewStandard, 1).SetBit(0, 1)
 
-	s := NewServer()
+	s := test.NewServer()
 	defer s.Close()
 	s.Handler.Host = s.Host()
-	s.Handler.Cluster = NewCluster(1)
+	s.Handler.Cluster = test.NewCluster(1)
 	s.Handler.Cluster.Nodes[0].Host = s.Host()
 	s.Handler.Holder = hldr.Holder
