+++ executor_test.go
@@ -25,12 +25,13 @@ import (
 	"github.com/davecgh/go-spew/spew"
 	"github.com/pilosa/pilosa"
 	"github.com/pilosa/pilosa/pql"
+	"github.com/pilosa/pilosa/test"
 )
 
 // Ensure a bitmap query can be executed.
 func TestExecutor_Execute_Bitmap(t *testing.T) {
 	t.Run("Row", func(t *testing.T) {
-		hldr := MustOpenHolder()
+		hldr := test.MustOpenHolder()
 		defer hldr.Close()
 		index := hldr.MustCreateIndexIfNotExists("i", pilosa.IndexOptions{})
 		f, err := index.CreateFrame("f", pilosa.FrameOptions{InverseEnabled: true})
@@ -38,7 +39,7 @@ func TestExecutor_Execute_Bitmap(t *testing.T) {
 			t.Fatal(err)
 		}
 
-		e := NewExecutor(hldr.Holder, NewCluster(1))
+		e := NewExecutor(hldr.Holder, test.NewCluster(1))
 
 		// Set bits.
 		if _, err := e.Execute(context.Background(), "i", MustParse(``+
@@ -62,14 +63,14 @@ func TestExecutor_Execute_Bitmap(t *testing.T) {
 	})
 
 	t.Run("Column", func(t *testing.T) {
-		hldr := MustOpenHolder()
+		hldr := test.MustOpenHolder()
 		defer hldr.Close()
 		index := hldr.MustCreateIndexIfNotExists("i", pilosa.IndexOptions{})
 		if _, err := index.CreateFrame("f", pilosa.FrameOptions{InverseEnabled: true}); err != nil {
 			t.Fatal(err)
 		}
 
-		e := NewExecutor(hldr.Holder, NewCluster(1))
+		e := NewExecutor(hldr.Holder, test.NewCluster(1))
 
 		// Set bits.
 		if _, err := e.Execute(context.Background(), "i", MustParse(``+
@@ -95,7 +96,7 @@ func TestExecutor_Execute_Bitmap(t *testing.T) {
 
 // Ensure a difference query can be executed.
 func TestExecutor_Execute_Difference(t *testing.T) {
-	hldr := MustOpenHolder()
+	hldr := test.MustOpenHolder()
 	defer hldr.Close()
 	hldr.MustCreateFragmentIfNotExists("i", "general", pilosa.ViewStandard, 0).MustSetBits(10, 1)
 	hldr.MustCreateFragmentIfNotExists("i", "general", pilosa.ViewStandard, 0).MustSetBits(10, 2)
@@ -103,7 +104,7 @@ func TestExecutor_Execute_Difference(t *testing.T) {
 	hldr.MustCreateFragmentIfNotExists("i", "general", pilosa.ViewStandard, 0).MustSetBits(11, 2)
 	hldr.MustCreateFragmentIfNotExists("i", "general", pilosa.ViewStandard, 0).MustSetBits(11, 4)
 
-	e := NewExecutor(hldr.Holder, NewCluster(1))
+	e := NewExecutor(hldr.Holder, test.NewCluster(1))
 	if res, err := e.Execute(context.Background(), "i", MustParse(`Difference(Bitmap(rowID=10), Bitmap(rowID=11))`), nil, nil); err != nil {
 		t.Fatal(err)
 	} else if bits := res[0].(*pilosa.Bitmap).Bits(); !reflect.DeepEqual(bits, []uint64{1, 3}) {
@@ -113,11 +114,11 @@ func TestExecutor_Execute_Difference(t *testing.T) {
 
 // Ensure an empty difference query behaves properly.
 func TestExecutor_Execute_Empty_Difference(t *testing.T) {
-	hldr := MustOpenHolder()
+	hldr := test.MustOpenHolder()
 	defer hldr.Close()
 	hldr.MustCreateFragmentIfNotExists("i", "general", pilosa.ViewStandard, 0).MustSetBits(10, 1)
 
-	e := NewExecutor(hldr.Holder, NewCluster(1))
+	e := NewExecutor(hldr.Holder, test.NewCluster(1))
 	if res, err := e.Execute(context.Background(), "i", MustParse(`Difference()`), nil, nil); err == nil {
 		t.Fatalf("Empty Difference query should give error, but got %v", res)
 	}
@@ -125,7 +126,7 @@ func TestExecutor_Execute_Empty_Difference(t *testing.T) {
 
 // Ensure an intersect query can be executed.
 func TestExecutor_Execute_Intersect(t *testing.T) {
-	hldr := MustOpenHolder()
+	hldr := test.MustOpenHolder()
 	defer hldr.Close()
 	hldr.MustCreateFragmentIfNotExists("i", "general", pilosa.ViewStandard, 0).MustSetBits(10, 1)
 	hldr.MustCreateFragmentIfNotExists("i", "general", pilosa.ViewStandard, 1).MustSetBits(10, SliceWidth+1)
@@ -135,7 +136,7 @@ func TestExecutor_Execute_Intersect(t *testing.T) {
 	hldr.MustCreateFragmentIfNotExists("i", "general", pilosa.ViewStandard, 0).MustSetBits(11, 2)
 	hldr.MustCreateFragmentIfNotExists("i", "general", pilosa.ViewStandard, 1).MustSetBits(11, SliceWidth+2)
 
-	e := NewExecutor(hldr.Holder, NewCluster(1))
+	e := NewExecutor(hldr.Holder, test.NewCluster(1))
 	if res, err := e.Execute(context.Background(), "i", MustParse(`Intersect(Bitmap(rowID=10), Bitmap(rowID=11))`), nil, nil); err != nil {
 		t.Fatal(err)
 	} else if bits := res[0].(*pilosa.Bitmap).Bits(); !reflect.DeepEqual(bits, []uint64{1, SliceWidth + 2}) {
@@ -145,10 +146,10 @@ func TestExecutor_Execute_Intersect(t *testing.T) {
 
 // Ensure an empty intersect query behaves properly.
 func TestExecutor_Execute_Empty_Intersect(t *testing.T) {
-	hldr := MustOpenHolder()
+	hldr := test.MustOpenHolder()
 	defer hldr.Close()
 
-	e := NewExecutor(hldr.Holder, NewCluster(1))
+	e := NewExecutor(hldr.Holder, test.NewCluster(1))
 	if res, err := e.Execute(context.Background(), "i", MustParse(`Intersect()`), nil, nil); err == nil {
 		t.Fatalf("Empty Intersect query should give error, but got %v", res)
 	}
@@ -156,7 +157,7 @@ func TestExecutor_Execute_Empty_Intersect(t *testing.T) {
 
 // Ensure a union query can be executed.
 func TestExecutor_Execute_Union(t *testing.T) {
-	hldr := MustOpenHolder()
+	hldr := test.MustOpenHolder()
 	defer hldr.Close()
 	hldr.MustCreateFragmentIfNotExists("i", "general", pilosa.ViewStandard, 0).MustSetBits(10, 0)
 	hldr.MustCreateFragmentIfNotExists("i", "general", pilosa.ViewStandard, 1).MustSetBits(10, SliceWidth+1)
@@ -165,7 +166,7 @@ func TestExecutor_Execute_Union(t *testing.T) {
 	hldr.MustCreateFragmentIfNotExists("i", "general", pilosa.ViewStandard, 0).MustSetBits(11, 2)
 	hldr.MustCreateFragmentIfNotExists("i", "general", pilosa.ViewStandard, 1).MustSetBits(11, SliceWidth+2)
 
-	e := NewExecutor(hldr.Holder, NewCluster(1))
+	e := NewExecutor(hldr.Holder, test.NewCluster(1))
 	if res, err := e.Execute(context.Background(), "i", MustParse(`Union(Bitmap(rowID=10), Bitmap(rowID=11))`), nil, nil); err != nil {
 		t.Fatal(err)
 	} else if bits := res[0].(*pilosa.Bitmap).Bits(); !reflect.DeepEqual(bits, []uint64{0, 2, SliceWidth + 1, SliceWidth + 2}) {
@@ -175,11 +176,11 @@ func TestExecutor_Execute_Union(t *testing.T) {
 
 // Ensure an empty union query behaves properly.
 func TestExecutor_Execute_Empty_Union(t *testing.T) {
-	hldr := MustOpenHolder()
+	hldr := test.MustOpenHolder()
 	defer hldr.Close()
 	hldr.MustCreateFragmentIfNotExists("i", "general", pilosa.ViewStandard, 0).MustSetBits(10, 0)
 
-	e := NewExecutor(hldr.Holder, NewCluster(1))
+	e := NewExecutor(hldr.Holder, test.NewCluster(1))
 	if res, err := e.Execute(context.Background(), "i", MustParse(`Union()`), nil, nil); err != nil {
 		t.Fatal(err)
 	} else if bits := res[0].(*pilosa.Bitmap).Bits(); !reflect.DeepEqual(bits, []uint64{}) {
@@ -189,13 +190,13 @@ func TestExecutor_Execute_Empty_Union(t *testing.T) {
 
 // Ensure a count query can be executed.
 func TestExecutor_Execute_Count(t *testing.T) {
-	hldr := MustOpenHolder()
+	hldr := test.MustOpenHolder()
 	defer hldr.Close()
 	hldr.MustCreateFragmentIfNotExists("i", "f", pilosa.ViewStandard, 0).MustSetBits(10, 3)
 	hldr.MustCreateFragmentIfNotExists("i", "f", pilosa.ViewStandard, 1).MustSetBits(10, SliceWidth+1)
 	hldr.MustCreateFragmentIfNotExists("i", "f", pilosa.ViewStandard, 1).MustSetBits(10, SliceWidth+2)
 
-	e := NewExecutor(hldr.Holder, NewCluster(1))
+	e := NewExecutor(hldr.Holder, test.NewCluster(1))
 	if res, err := e.Execute(context.Background(), "i", MustParse(`Count(Bitmap(rowID=10, frame=f))`), nil, nil); err != nil {
 		t.Fatal(err)
 	} else if res[0] != uint64(3) {
@@ -205,10 +206,10 @@ func TestExecutor_Execute_Count(t *testing.T) {
 
 // Ensure a set query can be executed.
 func TestExecutor_Execute_SetBit(t *testing.T) {
-	hldr := MustOpenHolder()
+	hldr := test.MustOpenHolder()
 	defer hldr.Close()
 
-	e := NewExecutor(hldr.Holder, NewCluster(1))
+	e := NewExecutor(hldr.Holder, test.NewCluster(1))
 	f := hldr.MustCreateFragmentIfNotExists("i", "f", pilosa.ViewStandard, 0)
 	if n := f.Row(11).Count(); n != 0 {
 		t.Fatalf("unexpected bitmap count: %d", n)
@@ -236,7 +237,7 @@ func TestExecutor_Execute_SetBit(t *testing.T) {
 
 // Ensure a SetRowAttrs() query can be executed.
 func TestExecutor_Execute_SetRowAttrs(t *testing.T) {
-	hldr := MustOpenHolder()
+	hldr := test.MustOpenHolder()
 	defer hldr.Close()
 
 	// Create frames.
@@ -249,7 +250,7 @@ func TestExecutor_Execute_SetRowAttrs(t *testing.T) {
 
 	// Set two fields on f/10.
 	// Also set fields on other bitmaps and frames to test isolation.
-	e := NewExecutor(hldr.Holder, NewCluster(1))
+	e := NewExecutor(hldr.Holder, test.NewCluster(1))
 	if _, err := e.Execute(context.Background(), "i", MustParse(`SetRowAttrs(rowID=10, frame=f, foo="bar")`), nil, nil); err != nil {
 		t.Fatal(err)
 	}
@@ -273,9 +274,9 @@ func TestExecutor_Execute_SetRowAttrs(t *testing.T) {
 
 // Ensure a TopN() query can be executed.
 func TestExecutor_Execute_TopN(t *testing.T) {
-	hldr := MustOpenHolder()
+	hldr := test.MustOpenHolder()
 	defer hldr.Close()
-	e := NewExecutor(hldr.Holder, NewCluster(1))
+	e := NewExecutor(hldr.Holder, test.NewCluster(1))
 
 	// Set bits for rows 0, 10, & 20 across two slices.
 	if idx, err := hldr.CreateIndex("i", pilosa.IndexOptions{}); err != nil {
@@ -326,7 +327,7 @@ func TestExecutor_Execute_TopN(t *testing.T) {
 	})
 }
 func TestExecutor_Execute_TopN_fill(t *testing.T) {
-	hldr := MustOpenHolder()
+	hldr := test.MustOpenHolder()
 	defer hldr.Close()
 
 	// Set bits for rows 0, 10, & 20 across two slices.
@@ -338,7 +339,7 @@ func TestExecutor_Execute_TopN_fill(t *testing.T) {
 	hldr.MustCreateRankedFragmentIfNotExists("i", "f", pilosa.ViewStandard, 1).SetBit(1, SliceWidth)
 
 	// Execute query.
-	e := NewExecutor(hldr.Holder, NewCluster(1))
+	e := NewExecutor(hldr.Holder, test.NewCluster(1))
 	if result, err := e.Execute(context.Background(), "i", MustParse(`TopN(frame=f, n=1)`), nil, nil); err != nil {
 		t.Fatal(err)
 	} else if !reflect.DeepEqual(result, []interface{}{[]pilosa.Pair{
@@ -350,7 +351,7 @@ func TestExecutor_Execute_TopN_fill(t *testing.T) {
 
 // Ensure
 func TestExecutor_Execute_TopN_fill_small(t *testing.T) {
-	hldr := MustOpenHolder()
+	hldr := test.MustOpenHolder()
 	defer hldr.Close()
 
 	hldr.MustCreateRankedFragmentIfNotExists("i", "f", pilosa.ViewStandard, 0).SetBit(0, 0)
@@ -372,7 +373,7 @@ func TestExecutor_Execute_TopN_fill_small(t *testing.T) {
 	hldr.MustCreateRankedFragmentIfNotExists("i", "f", pilosa.ViewStandard, 3).SetBit(4, 3*SliceWidth+1)
 
 	// Execute query.
-	e := NewExecutor(hldr.Holder, NewCluster(1))
+	e := NewExecutor(hldr.Holder, test.NewCluster(1))
 	if result, err := e.Execute(context.Background(), "i", MustParse(`TopN(frame=f, n=1)`), nil, nil); err != nil {
 		t.Fatal(err)
 	} else if !reflect.DeepEqual(result, []interface{}{[]pilosa.Pair{
@@ -384,7 +385,7 @@ func TestExecutor_Execute_TopN_fill_small(t *testing.T) {
 
 // Ensure a TopN() query with a source bitmap can be executed.
 func TestExecutor_Execute_TopN_Src(t *testing.T) {
-	hldr := MustOpenHolder()
+	hldr := test.MustOpenHolder()
 	defer hldr.Close()
 
 	// Set bits for rows 0, 10, & 20 across two slices.
@@ -407,7 +408,7 @@ func TestExecutor_Execute_TopN_Src(t *testing.T) {
 	hldr.MustCreateRankedFragmentIfNotExists("i", "other", pilosa.ViewStandard, 1).RecalculateCache()
 
 	// Execute query.
-	e := NewExecutor(hldr.Holder, NewCluster(1))
+	e := NewExecutor(hldr.Holder, test.NewCluster(1))
 	if result, err := e.Execute(context.Background(), "i", MustParse(`TopN(Bitmap(rowID=100, frame=other), frame=f, n=3)`), nil, nil); err != nil {
 		t.Fatal(err)
 	} else if !reflect.DeepEqual(result, []interface{}{[]pilosa.Pair{
@@ -422,7 +423,7 @@ func TestExecutor_Execute_TopN_Src(t *testing.T) {
 //Ensure TopN handles Attribute filters
 func TestExecutor_Execute_TopN_Attr(t *testing.T) {
 	//
-	hldr := MustOpenHolder()
+	hldr := test.MustOpenHolder()
 	defer hldr.Close()
 	hldr.MustCreateRankedFragmentIfNotExists("i", "f", pilosa.ViewStandard, 0).SetBit(0, 0)
 	hldr.MustCreateRankedFragmentIfNotExists("i", "f", pilosa.ViewStandard, 0).SetBit(0, 1)
@@ -431,7 +432,7 @@ func TestExecutor_Execute_TopN_Attr(t *testing.T) {
 	if err := hldr.Frame("i", "f").RowAttrStore().SetAttrs(10, map[string]interface{}{"category": int64(123)}); err != nil {
 		t.Fatal(err)
 	}
-	e := NewExecutor(hldr.Holder, NewCluster(1))
+	e := NewExecutor(hldr.Holder, test.NewCluster(1))
 	if result, err := e.Execute(context.Background(), "i", MustParse(`TopN(frame="f", n=1, field="category", filters=[123])`), nil, nil); err != nil {
 		t.Fatal(err)
 	} else if !reflect.DeepEqual(result, []interface{}{[]pilosa.Pair{
@@ -445,7 +446,7 @@ func TestExecutor_Execute_TopN_Attr(t *testing.T) {
 //Ensure TopN handles Attribute filters with source bitmap
 func TestExecutor_Execute_TopN_Attr_Src(t *testing.T) {
 	//
-	hldr := MustOpenHolder()
+	hldr := test.MustOpenHolder()
 	defer hldr.Close()
 	hldr.MustCreateRankedFragmentIfNotExists("i", "f", pilosa.ViewStandard, 0).SetBit(0, 0)
 	hldr.MustCreateRankedFragmentIfNotExists("i", "f", pilosa.ViewStandard, 0).SetBit(0, 1)
@@ -454,7 +455,7 @@ func TestExecutor_Execute_TopN_Attr_Src(t *testing.T) {
 	if err := hldr.Frame("i", "f").RowAttrStore().SetAttrs(10, map[string]interface{}{"category": uint64(123)}); err != nil {
 		t.Fatal(err)
 	}
-	e := NewExecutor(hldr.Holder, NewCluster(1))
+	e := NewExecutor(hldr.Holder, test.NewCluster(1))
 	if result, err := e.Execute(context.Background(), "i", MustParse(`TopN(Bitmap(rowID=10,frame=f),frame="f", n=1, field="category", filters=[123])`), nil, nil); err != nil {
 		t.Fatal(err)
 	} else if !reflect.DeepEqual(result, []interface{}{[]pilosa.Pair{
@@ -467,9 +468,9 @@ func TestExecutor_Execute_TopN_Attr_Src(t *testing.T) {
 
 // Ensure a range query can be executed.
 func TestExecutor_Execute_Range(t *testing.T) {
-	hldr := MustOpenHolder()
+	hldr := test.MustOpenHolder()
 	defer hldr.Close()
-	e := NewExecutor(hldr.Holder, NewCluster(1))
+	e := NewExecutor(hldr.Holder, test.NewCluster(1))
 
 	// Create index.
 	index := hldr.MustCreateIndexIfNotExists("i", pilosa.IndexOptions{})
@@ -507,7 +508,7 @@ func TestExecutor_Execute_Range(t *testing.T) {
 	})
 
 	t.Run("Inverse", func(t *testing.T) {
-		e := NewExecutor(hldr.Holder, NewCluster(1))
+		e := NewExecutor(hldr.Holder, test.NewCluster(1))
 		if res, err := e.Execute(context.Background(), "i", MustParse(`Range(columnID=2, frame=f, start="1999-01-01T00:00", end="2003-01-01T00:00")`), nil, nil); err != nil {
 			t.Fatal(err)
 		} else if bits := res[0].(*pilosa.Bitmap).Bits(); !reflect.DeepEqual(bits, []uint64{1, 10}) {
@@ -518,10 +519,10 @@ func TestExecutor_Execute_Range(t *testing.T) {
 
 // Ensure a remote query can return a bitmap.
 func TestExecutor_Execute_Remote_Bitmap(t *testing.T) {
-	c := NewCluster(2)
+	c := test.NewCluster(2)
 
 	// Create secondary server and update second cluster node.
-	s := NewServer()
+	s := test.NewServer()
 	defer s.Close()
 	c.Nodes[1].Host = s.Host()
 
@@ -546,7 +547,7 @@ func TestExecutor_Execute_Remote_Bitmap(t *testing.T) {
 
 	// Create local executor data.
 	// The local node owns slice 1.
-	hldr := MustOpenHolder()
+	hldr := test.MustOpenHolder()
 	defer hldr.Close()
 	s.Handler.Holder = hldr.Holder
 	hldr.MustCreateFragmentIfNotExists("i", "f", pilosa.ViewStandard, 1).MustSetBits(10, (1*SliceWidth)+1)
@@ -561,10 +562,10 @@ func TestExecutor_Execute_Remote_Bitmap(t *testing.T) {
 
 // Ensure a remote query can return a count.
 func TestExecutor_Execute_Remote_Count(t *testing.T) {
-	c := NewCluster(2)
+	c := test.NewCluster(2)
 
 	// Create secondary server and update second cluster node.
-	s := NewServer()
+	s := test.NewServer()
 	defer s.Close()
 	c.Nodes[1].Host = s.Host()
 
@@ -574,7 +575,7 @@ func TestExecutor_Execute_Remote_Count(t *testing.T) {
 	}
 
 	// Create local executor data. The local node owns slice 1.
-	hldr := MustOpenHolder()
+	hldr := test.MustOpenHolder()
 	defer hldr.Close()
 	s.Handler.Holder = hldr.Holder
 	hldr.MustCreateFragmentIfNotExists("i", "f", pilosa.ViewStandard, 2).MustSetBits(10, (2*SliceWidth)+1)
@@ -590,11 +591,11 @@ func TestExecutor_Execute_Remote_Count(t *testing.T) {
 
 // Ensure a remote query can set bits on multiple nodes.
 func TestExecutor_Execute_Remote_SetBit(t *testing.T) {
-	c := NewCluster(2)
+	c := test.NewCluster(2)
 	c.ReplicaN = 2
 
 	// Create secondary server and update second cluster node.
-	s := NewServer()
+	s := test.NewServer()
 	defer s.Close()
 	c.Nodes[1].Host = s.Host()
 
@@ -611,7 +612,7 @@ func TestExecutor_Execute_Remote_SetBit(t *testing.T) {
 	}
 
 	// Create local executor data.
-	hldr := MustOpenHolder()
+	hldr := test.MustOpenHolder()
 	defer hldr.Close()
 	s.Handler.Holder = hldr.Holder
 
@@ -636,11 +637,11 @@ func TestExecutor_Execute_Remote_SetBit(t *testing.T) {
 
 // Ensure a remote query can set bits on multiple nodes.
 func TestExecutor_Execute_Remote_SetBit_With_Timestamp(t *testing.T) {
-	c := NewCluster(2)
+	c := test.NewCluster(2)
 	c.ReplicaN = 2
 
 	// Create secondary server and update second cluster node.
-	s := NewServer()
+	s := test.NewServer()
 	defer s.Close()
 	c.Nodes[1].Host = s.Host()
 
@@ -657,7 +658,7 @@ func TestExecutor_Execute_Remote_SetBit_With_Timestamp(t *testing.T) {
 	}
 
 	// Create local executor data.
-	hldr := MustOpenHolder()
+	hldr := test.MustOpenHolder()
 	defer hldr.Close()
 	s.Handler.Holder = hldr.Holder
 
@@ -684,10 +685,10 @@ func TestExecutor_Execute_Remote_SetBit_With_Timestamp(t *testing.T) {
 
 // Ensure a remote query can return a top-n query.
 func TestExecutor_Execute_Remote_TopN(t *testing.T) {
-	c := NewCluster(2)
+	c := test.NewCluster(2)
 
 	// Create secondary server and update second cluster node.
-	s := NewServer()
+	s := test.NewServer()
 	defer s.Close()
 	c.Nodes[1].Host = s.Host()
 
@@ -725,7 +726,7 @@ func TestExecutor_Execute_Remote_TopN(t *testing.T) {
 	}
 
 	// Create local executor data on slice 2 & 4.
-	hldr := MustOpenHolder()
+	hldr := test.MustOpenHolder()
 	defer hldr.Close()
 	s.Handler.Holder = hldr.Holder
 	hldr.MustCreateRankedFragmentIfNotExists("i", "f", pilosa.ViewStandard, 2).MustSetBits(30, (2*SliceWidth)+1)
@@ -745,9 +746,9 @@ func TestExecutor_Execute_Remote_TopN(t *testing.T) {
 
 // Ensure executor returns an error if too many writes are in a single request.
 func TestExecutor_Execute_ErrMaxWritesPerRequest(t *testing.T) {
-	hldr := MustOpenHolder()
+	hldr := test.MustOpenHolder()
 	defer hldr.Close()
-	e := NewExecutor(hldr.Holder, NewCluster(1))
+	e := NewExecutor(hldr.Holder, test.NewCluster(1))
 	e.MaxWritesPerRequest = 3
 	if _, err := e.Execute(context.Background(), "i", MustParse(`SetBit() ClearBit() SetBit() SetBit()`), nil, nil); err != pilosa.ErrTooManyWrites {
 		t.Fatalf("unexpected error: %s", err)
