+++ index_test.go
@@ -15,17 +15,16 @@
 package pilosa_test
 
 import (
-	"io/ioutil"
-	"os"
 	"reflect"
 	"testing"
 
 	"github.com/pilosa/pilosa"
+	"github.com/pilosa/pilosa/test"
 )
 
 // Ensure index can open and retrieve a frame.
 func TestIndex_CreateFrameIfNotExists(t *testing.T) {
-	index := MustOpenIndex()
+	index := test.MustOpenIndex()
 	defer index.Close()
 
 	// Create frame.
@@ -53,7 +52,7 @@ func TestIndex_CreateFrame(t *testing.T) {
 	// Ensure time quantum can be set appropriately on a new frame.
 	t.Run("TimeQuantum", func(t *testing.T) {
 		t.Run("Explicit", func(t *testing.T) {
-			index := MustOpenIndex()
+			index := test.MustOpenIndex()
 			defer index.Close()
 
 			// Set index time quantum.
@@ -71,7 +70,7 @@ func TestIndex_CreateFrame(t *testing.T) {
 		})
 
 		t.Run("Inherited", func(t *testing.T) {
-			index := MustOpenIndex()
+			index := test.MustOpenIndex()
 			defer index.Close()
 
 			// Set index time quantum.
@@ -92,7 +91,7 @@ func TestIndex_CreateFrame(t *testing.T) {
 	// Ensure frame can include range columns.
 	t.Run("RangeEnabled", func(t *testing.T) {
 		t.Run("OK", func(t *testing.T) {
-			index := MustOpenIndex()
+			index := test.MustOpenIndex()
 			defer index.Close()
 
 			// Create frame with schema and verify it exists.
@@ -127,7 +126,7 @@ func TestIndex_CreateFrame(t *testing.T) {
 		})
 
 		t.Run("ErrInverseRangeNotAllowed", func(t *testing.T) {
-			index := MustOpenIndex()
+			index := test.MustOpenIndex()
 			defer index.Close()
 
 			if _, err := index.CreateFrame("f", pilosa.FrameOptions{
@@ -139,7 +138,7 @@ func TestIndex_CreateFrame(t *testing.T) {
 		})
 
 		t.Run("ErrRangeCacheNotAllowed", func(t *testing.T) {
-			index := MustOpenIndex()
+			index := test.MustOpenIndex()
 			defer index.Close()
 
 			if _, err := index.CreateFrame("f", pilosa.FrameOptions{
@@ -151,7 +150,7 @@ func TestIndex_CreateFrame(t *testing.T) {
 		})
 
 		t.Run("ErrFrameFieldsNotAllowed", func(t *testing.T) {
-			index := MustOpenIndex()
+			index := test.MustOpenIndex()
 			defer index.Close()
 
 			if _, err := index.CreateFrame("f", pilosa.FrameOptions{
@@ -164,7 +163,7 @@ func TestIndex_CreateFrame(t *testing.T) {
 		})
 
 		t.Run("ErrFieldNameRequired", func(t *testing.T) {
-			index := MustOpenIndex()
+			index := test.MustOpenIndex()
 			defer index.Close()
 
 			if _, err := index.CreateFrame("f", pilosa.FrameOptions{
@@ -178,7 +177,7 @@ func TestIndex_CreateFrame(t *testing.T) {
 		})
 
 		t.Run("ErrInvalidFieldType", func(t *testing.T) {
-			index := MustOpenIndex()
+			index := test.MustOpenIndex()
 			defer index.Close()
 
 			if _, err := index.CreateFrame("f", pilosa.FrameOptions{
@@ -192,7 +191,7 @@ func TestIndex_CreateFrame(t *testing.T) {
 		})
 
 		t.Run("ErrInvalidFieldRange", func(t *testing.T) {
-			index := MustOpenIndex()
+			index := test.MustOpenIndex()
 			defer index.Close()
 
 			if _, err := index.CreateFrame("f", pilosa.FrameOptions{
@@ -209,7 +208,7 @@ func TestIndex_CreateFrame(t *testing.T) {
 	// Ensure frame cannot be created with a matching row label.
 	t.Run("ErrColumnRowLabelEqual", func(t *testing.T) {
 		t.Run("Explicit", func(t *testing.T) {
-			index := MustOpenIndex()
+			index := test.MustOpenIndex()
 			defer index.Close()
 
 			_, err := index.CreateFrame("f", pilosa.FrameOptions{RowLabel: pilosa.DefaultColumnLabel})
@@ -219,7 +218,7 @@ func TestIndex_CreateFrame(t *testing.T) {
 		})
 
 		t.Run("Default", func(t *testing.T) {
-			index := MustOpenIndex()
+			index := test.MustOpenIndex()
 			defer index.Close()
 			if err := index.SetColumnLabel(pilosa.DefaultRowLabel); err != nil {
 				t.Fatal(err)
@@ -235,7 +234,7 @@ func TestIndex_CreateFrame(t *testing.T) {
 
 // Ensure index can delete a frame.
 func TestIndex_DeleteFrame(t *testing.T) {
-	index := MustOpenIndex()
+	index := test.MustOpenIndex()
 	defer index.Close()
 
 	// Create frame.
@@ -258,7 +257,7 @@ func TestIndex_DeleteFrame(t *testing.T) {
 
 // Ensure index can set the default time quantum.
 func TestIndex_SetTimeQuantum(t *testing.T) {
-	index := MustOpenIndex()
+	index := test.MustOpenIndex()
 	defer index.Close()
 
 	// Set & retrieve time quantum.
@@ -273,87 +272,5 @@ func TestIndex_SetTimeQuantum(t *testing.T) {
 		t.Fatal(err)
 	} else if q := index.TimeQuantum(); q != pilosa.TimeQuantum("YMDH") {
 		t.Fatalf("unexpected quantum (reopen): %s", q)
-	}
-}
-
-// Index represents a test wrapper for pilosa.Index.
-type Index struct {
-	*pilosa.Index
-}
-
-// NewIndex returns a new instance of Index.
-func NewIndex() *Index {
-	path, err := ioutil.TempDir("", "pilosa-index-")
-	if err != nil {
-		panic(err)
-	}
-	index, err := pilosa.NewIndex(path, "i")
-	if err != nil {
-		panic(err)
-	}
-	return &Index{Index: index}
-}
-
-// MustOpenIndex returns a new, opened index at a temporary path. Panic on error.
-func MustOpenIndex() *Index {
-	index := NewIndex()
-	if err := index.Open(); err != nil {
-		panic(err)
-	}
-	return index
-}
-
-// Close closes the index and removes the underlying data.
-func (i *Index) Close() error {
-	defer os.RemoveAll(i.Path())
-	return i.Index.Close()
-}
-
-// Reopen closes the index and reopens it.
-func (i *Index) Reopen() error {
-	var err error
-	if err := i.Index.Close(); err != nil {
-		return err
-	}
-
-	path, name := i.Path(), i.Name()
-	i.Index, err = pilosa.NewIndex(path, name)
-	if err != nil {
-		return err
-	}
-
-	if err := i.Open(); err != nil {
-		return err
-	}
-	return nil
-}
-
-// CreateFrame creates a frame with the given options.
-func (i *Index) CreateFrame(name string, opt pilosa.FrameOptions) (*Frame, error) {
-	f, err := i.Index.CreateFrame(name, opt)
-	if err != nil {
-		return nil, err
-	}
-	return &Frame{Frame: f}, nil
-}
-
-// CreateFrameIfNotExists creates a frame with the given options if it doesn't exist.
-func (i *Index) CreateFrameIfNotExists(name string, opt pilosa.FrameOptions) (*Frame, error) {
-	f, err := i.Index.CreateFrameIfNotExists(name, opt)
-	if err != nil {
-		return nil, err
-	}
-	return &Frame{Frame: f}, nil
-}
-
-// Ensure index can delete a frame.
-func TestIndex_InvalidName(t *testing.T) {
-	path, err := ioutil.TempDir("", "pilosa-index-")
-	if err != nil {
-		panic(err)
-	}
-	index, err := pilosa.NewIndex(path, "ABC")
-	if index != nil {
-		t.Fatalf("unexpected index name %s", index)
 	}
 }
